Moved external modules to dandi/support/*/
